More graphicaudio.sh improvements
* Require passing author in to avoid mistakes (author not extractable from file)
* Replace colon in titles as it is unsafe on several filesystems
* Improved error checking/handling

+++ graphicaudio.sh
@@ -6,11 +6,23 @@ shopt -qs extglob
 # Graphic Audio is already DRM-free, and the files large enough that it rarely
 # makes sense to combine them. We only need to unzip and rename.
 
-if [[ $# -lt 1 ]]; then
-  echo "Usage: ${0} GRAPHIC_AUDIO_ZIPFILES..." 1>&2
+function usage {
+  # Graphic Audio does not cleanly embed author name in their books - have to pass manually
+  echo "Usage: ${0} 'AUTHOR' GRAPHIC_AUDIO_ZIPFILES..." 1>&2
+}
+
+if [[ $# -lt 2 ]]; then
+  usage
   exit 1
 fi
 
+if [[ -f "$1" ]]; then
+  echo "ERROR: First argument should be quoted author name, but it looks like a file path, aborting" 1>&2
+  usage
+  exit 3
+fi
+
+# TODO: Extract output paths to a common file for all scripts
 if [[ -d /mnt/dropbox/Archive/Audiobooks ]]; then
   readonly output="/mnt/dropbox/Archive/Audiobooks/Graphic Audio"
 elif [[ -d /Volumes/dropbox/Archive/Audiobooks ]]; then
@@ -20,17 +32,21 @@ else
   exit 2
 fi
 
-# Graphic Audio does not cleanly embed author name in their books - have to set manually
-author="Brent Weeks"
+author="$1"
+shift 1
 
 for zipfile in $@; do
+  if [[ ! -f "$zipfile" ]]; then
+    echo "'$zipfile' not found! Aborting" 1>&2
+    exit 4
+  fi
   filename="$(zipinfo -1 "$zipfile" | grep -E '\.m4b$')"
   unzip "$zipfile" -d /tmp
   echo "FILENAME: $filename" 1>&2
   (
     cd /tmp
     title="$(ffprobe -v error -show_entries format_tags=album -of default=noprint_wrappers=1:nokey=1 "$filename" \
-      | perl -pe 's/(\d of \d)/Graphic Audio \1/g')"
+      | perl -pe 's/(\d of \d)/Graphic Audio \1/;s/:/ -/g')"
     output_path="${output}/${title} - ${author}.m4b"
     echo "Copying M4B file to archival: ${output_path}" 1>&2
     mv "$filename" "${output_path}"
